common/objects.c: Use calloc() instead of malloc() + manual NULL'ing
calloc() always returns a fully null'ed out buffer. In case we
get the memory from the kernel, this has already been done, so
the manual setting of variables was quite unnecessary. As a
nice side-effect, we save about 250 lines of code, some random
cycles and can get rid of even more conditional code.

Signed-off-by: Andreas Ericsson <ae@op5.se>

Author: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1485 5f96b256-904b-4d8d-8c98-d829582c6739