base/utils: Refactor my_fcopy()
This patch provides some much-needed fixing to my_fcopy() by a
more or less complete rewrite of the function.

Previously it didn't use to check that it had, indeed, written
all data to the destination file. This is now done.

It also used a ridiculously small buffer of 1K, which is clearly
suboptimal for the task. A value of 128K or filesize(source),
whichever is smallest, should allow us to utilize the disk's
cache better without being so large that it drops out of the L2
cache.

It would also ignore the error conditions of "EINTR" and "EAGAIN",
which can happen if we receive a signal in the middle of reading
or writing, and, on failure after having opened the destination
file, it would leave the half-written (or empty) file around,
which could cause a major build-up of checkresult files in the
spool directory after some time.

It was triggered by the fact that my_fcopy() used to check the
return code of open() wrong, and spiralled slightly out of
control from there.

All these issues are addressed with the new code, which also
uses the "check for error and return if so" style of coding to
prevent indentation going to 5 levels deep or more.

Breaking it into two separate functions also makes it usable
when race-lessly creating temporary files using functions
like mkstemp().

Signed-off-by: Andreas Ericsson <ae@op5.se>

Author: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1504 5f96b256-904b-4d8d-8c98-d829582c6739