core: expect UTF-8 strings from the monitoring engine.
Internally, we use the QString object from Qt which encodes strings
in UTF-16. Strings that were fetched from the monitoring engine
were decoded using a Latin1 codec which lead to undesirable results
were strings were encoded as UTF-8 (which was the intended
encoding). Now, all exchange with the outside world (from is the
monitoring engine and to are the BBDO and NDO protocols) are made
using a propre UTF-8 encoding whereas internal strings keep being
encoded with UTF-16.