SQL: properly process /command_line/ field of hosts and services tables.
This issue was caused by a latency issue between the time at which
the command line was generated and the time at which it was processed
by the SQL endpoint (it should be less than 1 second). There is now
a much longer acceptable delay of 5 minutes that ough to be enough
for anybody. This fixes #5810.