BAM: fix BA and KPI events times.
Previously BA events were created in response to a change of the
virtual service of the BA (which itself reacted to a database
change introduced by the BAM module). However this created erronous
timing information in BA events. Now BA and KPI events start/end
times are solely based on KPI changing times.