ci(Gorgone): add firsts legacycmd tests
Check gorgone retrieve data from the /var/lib/centreon/centcore/ directory, and send the forced check to the engine pipe.
We use a perl script to mock the pipe, as we don't want to run engine, and gorgone check the file is a real pipe and not a simple file.
Check gorgone push engine, broken and vmware configuration files to the poller, and the poller set the correct path.
Gorgone don't seem to correctly take the configuration modules:engine:command_file. The default /var/lib/centreon-engine/rw/centengine.cmd is used instead, this should be investigated in a separate ticket.
gorgone don't set user/group when processing folder with untar.
create centcore folder in each tests
As www-data don't exist on rpm based distro, we can't test if user/group is correctly set by gorgone.

Refs:MON-144643