fix(gorgone): pullwss split GETLOG message before sending them to another host if needed.
- getlog() always order select query by id.
- pullwss module split any GETLOG message to stay bellow the websocket message size limit and transform them in SETLOGS.

For now if a message is lost in transmission it will never be asked again for. An error message will be present in the central log file.

The split is made in the pullwss and not in send_internal_action because the http api use the same GETLOG message locally, which would be broken if messages are split. the new behaviour allow the api to send back as many logs as needed, and to split log only for pullwss connexion method. The other connexion method use zmq that don't have a size limit.

MON-157180