wip(engine): Difference application from configuration sent by broker well advanced.
We have an issue with objects without keys like hostdependencies or hostescalations. Engine allows
them to be defined in a first time for many hosts/services and expands this definition later.
But as the used key to store them depends of the content, it is difficult to get them back after
expansion.
We have to think about to store them in a set, but there is an impact with some already written code.