fix(packaging): Add perl(Tie::File) dependency in gorgone rpm package (#2300)
On Alma9, this module has been split out and needs to be explicitly installed.
On Alma8, it worked fine as it is part of perl-interpreter -- but the new
dependency will not be a problem as perl-interpreter provides it.

On Debian 11 and 12, the Tie::File perl module is part of the perl-modules
packages; no need to do anything.